requirements.txt
altgraph==0.16.1
asn1crypto==1.3.0
atomicwrites==1.4.0
attrs==19.3.0
auto-py-to-exe==2.7.3
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
bottle==0.12.18
bottle-websocket==0.2.9
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
chromedriver-binary==83.0.4103.39.0
click==7.1.2
colorama==0.4.3
configparser==4.0.2
constantly==15.1.0
cryptography==2.3.1
cssselect==1.1.0
demjson==2.2.4
Eel==0.11.0
Flask==1.1.2
future==0.18.2
gevent==20.5.0
gevent-websocket==0.10.1
google-api-core==1.21.0
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
greenlet==0.4.15
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.1.1
macholib==1.11
MarkupSafe==1.1.1
more-itertools==8.2.0
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.3
pandas==1.0.3
parsel==1.5.2
pefile==2019.4.18
pluggy==0.13.1
pprint==0.1
protobuf==3.12.2
py==1.8.1
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pycrypto==2.6.1
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyInstaller==3.6
pyOpenSSL==19.0.0
pyparsing==2.4.7
PySimpleGUI==4.19.0
PySocks==1.7.1
pytest==5.4.1
pytest-runner==5.2
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
Twisted==20.3.0
urllib3==1.25.8
w3lib==1.21.0
wcwidth==0.1.9
Werkzeug==1.0.1
whichcraft==0.6.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.1.0
zope.interface==4.7.1
